+++ Backend.Handler/Services/OMDb/IOMDbService.cs
@@ -5,6 +5,6 @@ namespace Backend.Handler.Services.OMDb
 {
     public interface IOMDbService
     {
-        Task<OMDbResponse> RequestMoviesAsync(string title, int? year, byte plot);
+        Task<OMDbResponse> RequestMoviesAsync(string title, int? year, int plot);
     }
 }
